+++ actix-http/CHANGES.md
@@ -7,10 +7,11 @@
 * `AnyBody::into_boxed` for quickly converting to a type-erased, boxed body type. [#????]
 
 ### Changed
-* Rename `AnyBody::{Message => Stream}`. [#2446]
+* Rename `AnyBody::{Message => Body}`. [#2446]
 * Rename `AnyBody::{from_message => new_boxed}`. [#????]
+* Rename `AnyBody::{from_slice => copy_from_slice}`. [#????]
 * Rename `BoxAnyBody` to `BoxBody` [#????]
-* Change representation of `AnyBody` to include a type parameter in `Stream` variant. Defaults to `BoxBody`. [#????]
+* Change representation of `AnyBody` to include a type parameter in `Body` variant. Defaults to `BoxBody`. [#????]
 
 ### Removed
 * `AnyBody::Empty`; an empty body can now only be represented as a zero-length `Bytes` variant. [#2446]